Adds Redux to manage state and dummy "back-end"
This commit adds Redux and adds a redux store to manage the state for
the list of projects in the portfolio.

This commit also creates a "service" to get portfolio/project data.
Because there is no back-end live anywhere, this just mimics the
behavior of what an actual Axios service might return, and instead
pulls data from a separate file ('/services/data/portfolioData.js').