Adds filter/search by keyword functionality
This commit adds a filter text entry which filters the project list
by keywords. The filter is managed in the Redux store, and has its
own reducer and associated functions.

When 10 or fewer keywords match the filter text, the app displays
buttons for each, and clicking on one will set the filter to show
exclusively projects matching that exact keyword. This is different
from typing in a keyword, as typed searches will include all matching
keywords, even if the text entered matches another keyword exactly
(e.g., if a user types in 'react', projects with either the 'react'
or 'react-css' keyword would be displayed; if the user clicks the
'react' button, only projects with the 'react' keyword will be
displayed).